Upgrade gradle, fix test logging and int test CPU usage
- Upgrade to gradle 5.2.1. As part of this, I have separated the
  integration tests into their own task so that they can be run
  independently. The "allTest" target will do what the "test" target
  used to do.
- Suppress some especially noisy logs in integration tests. These log
  messages were firing continuously when the node-down scenarios were
  tested.
- Configure MockCluster replication threads to do a 100 ms sleep per
  iteration. This greatly reduces the CPU requirements for running the
  integration tests and allows us to run them on travis-ci.
- Set up travis to run integration tests too.
