+++ .travis.yml
@@ -1,7 +1,7 @@
 language: java
 dist: trusty
 script:
-  - ./gradlew build && ./gradlew codeCoverageReport
+  - ./gradlew build codeCoverageReport
 sudo: required
 jdk:
   - oraclejdk8
