+++ ambry-server/src/integration-test/java/com.github.ambry.server/MockCluster.java
@@ -57,10 +57,10 @@ import static org.junit.Assert.*;
  * On shutdown we ensure the servers are shutdown.
  */
 public class MockCluster {
-  private Logger logger = LoggerFactory.getLogger(getClass());
+  private static final Logger logger = LoggerFactory.getLogger(MockCluster.class);
   private final MockClusterAgentsFactory mockClusterAgentsFactory;
   private final MockClusterMap clusterMap;
-  private List<AmbryServer> serverList = null;
+  private final List<AmbryServer> serverList;
   private boolean serverInitialized = false;
   private int generalDataNodeIndex;
   private int prefetchDataNodeIndex;
@@ -88,7 +88,7 @@ public class MockCluster {
             numStoresPerMountPoint);
     clusterMap = mockClusterAgentsFactory.getClusterMap();
 
-    serverList = new ArrayList<AmbryServer>();
+    serverList = new ArrayList<>();
     generalDataNodeIndex = 0;
     prefetchDataNodeIndex = clusterMap.getDataNodes().size() - 1;
   }
@@ -152,6 +152,8 @@ public class MockCluster {
     props.setProperty("store.validate.authorization", "true");
     props.setProperty("kms.default.container.key", TestUtils.getRandomKey(32));
     props.setProperty("server.enable.store.data.prefetch", Boolean.toString(enableDataPrefetch));
+    props.setProperty("replication.intra.replica.thread.throttle.sleep.duration.ms", "100");
+    props.setProperty("replication.inter.replica.thread.throttle.sleep.duration.ms", "100");
     props.putAll(sslProperties);
     VerifiableProperties propverify = new VerifiableProperties(props);
     AmbryServer server = new AmbryServer(propverify, mockClusterAgentsFactory, notificationSystem, time);
