+++ build.gradle
@@ -39,14 +39,19 @@ subprojects {
     targetCompatibility = 1.8
 
     sourceSets {
-        test {
+        // separate source set for integration tests
+        intTest {
             java.srcDir file('src/integration-test/java')
             resources.srcDir file('src/integration-test/resources')
+            compileClasspath += sourceSets.main.output + sourceSets.test.output
+            runtimeClasspath += sourceSets.main.output + sourceSets.test.output
         }
-        unitTest {
-            java.srcDir file('src/test/java')
-            resources.srcDir file('src/test/resources')
-        }
+    }
+
+    configurations {
+        // integration tests should be able to get the same dependencies as the corresponding unit tests
+        intTestCompile.extendsFrom testCompile
+        intTestRuntime.extendsFrom testRuntime
     }
 
     dependencies {
@@ -55,11 +60,16 @@ subprojects {
         compile "org.slf4j:slf4j-log4j12:$slf4jVersion"
         testCompile "junit:junit:$junitVersion"
         testCompile "org.mockito:mockito-core:2.+"
-        unitTestCompile sourceSets.main.output
-        unitTestCompile sourceSets.test.output
-        unitTestCompile configurations.testCompile
-        unitTestRuntime configurations.testRuntime
         testRuntime project(':log4j-test-config')
+    }
+
+    idea {
+        // since integration tests are in a separate source set, we have to tell intellij about them
+        module {
+            testSourceDirs += sourceSets.intTest.java.srcDirs
+            testResourceDirs += sourceSets.intTest.resources.srcDirs
+            scopes.TEST.plus += [configurations.intTestCompile]
+        }
     }
 
     test {
@@ -68,18 +78,22 @@ subprojects {
         }
     }
 
-    task unitTest(type: Test) {
+    task intTest(type: Test) {
+        description = 'Runs integration tests.'
+        group = 'verification'
+        testClassesDirs = sourceSets.intTest.output.classesDirs
+        classpath = sourceSets.intTest.runtimeClasspath
         testLogging {
             exceptionFormat = 'full'
         }
-        testClassesDir = sourceSets.unitTest.output.classesDir
-        classpath += sourceSets.unitTest.runtimeClasspath
     }
 
-    // only run unit tests when doing a standard build,
-    // but run integration tests too when running the test target
-    check.dependsOn -= test
-    check.dependsOn += unitTest
+    task allTest {
+        description = 'Runs unit and integration tests'
+        group = 'verification'
+    }
+    allTest.dependsOn test
+    allTest.dependsOn intTest
 
     task testJar(type: Jar) {
         from sourceSets.test.output
@@ -377,12 +391,11 @@ task codeCoverageReport(type: JacocoReport) {
 
     reports {
         xml.enabled true
-        xml.destination "${buildDir}/reports/jacoco/report.xml"
+        xml.destination file("${buildDir}/reports/jacoco/report.xml")
         html.enabled true
         csv.enabled true
     }
 }
 
-codeCoverageReport.dependsOn {
-    subprojects*.unitTest
-}
+codeCoverageReport.dependsOn subprojects*.allTest
+
