+++ gradle.properties
@@ -12,4 +12,4 @@
 version=0.1.0
 org.gradle.daemon=true
 org.gradle.configureondemand=true
-org.gradle.jvmargs=-Xmx3g -XX:MaxPermSize=4096m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+org.gradle.jvmargs=-Xmx4g -XX:MaxPermSize=4096m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
