+++ gradle/maven.gradle
@@ -8,6 +8,7 @@
 // under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 // CONDITIONS OF ANY KIND, either express or implied.
 subprojects {
+    apply plugin: 'java'
     apply plugin: 'maven-publish'
 
     publishing {
