+++ log4j-test-config/src/main/resources/log4j.properties
@@ -33,3 +33,8 @@ log4j.additivity.PublicAccessLogger = false
 log4j.logger.org.apache.helix=WARN
 log4j.logger.org.apache.zookeeper=WARN
 log4j.logger.org.I0Itec.zkclient=WARN
+# these three are disabled because MockClusterMap does not currently mark nodes offline when requests fail, which
+# happens a lot in the cases that our end to end tests cover.
+log4j.logger.com.github.ambry.network.BlockingChannelConnectionPool=OFF
+log4j.logger.com.github.ambry.network.BlockingChannelInfo=OFF
+log4j.logger.com.github.ambry.replication.ReplicaThread=OFF
